+++ frontend/src/components/rec-resource/Card/CardCarousel.tsx
