+++ frontend/src/components/search-page/Search.scss
