+++ frontend/src/components/search-page/SearchBanner.tsx
@@ -1,7 +1,7 @@
 import { useMemo } from 'react';
 import { useSearchParams } from 'react-router-dom';
 import debounce from 'lodash.debounce';
-import '@/components/search/Search.scss';
+import '@/components/search-page/Search.scss';
 
 const SearchBanner = () => {
   const [searchParams, setSearchParams] = useSearchParams();
