+++ frontend/src/components/search-page/SearchPage.tsx
@@ -3,8 +3,8 @@ import { useSearchParams } from 'react-router-dom';
 import apiService from '@/service/api-service';
 import type { AxiosResponse } from '~/axios';
 import buildQueryString from '@/utils/buildQueryString';
-import RecResourceCard from '@/components/rec-resource/Card/RecResourceCard';
-import SearchBanner from '@/components/search/SearchBanner';
+import RecResourceCard from '@/components/rec-resource/card/RecResourceCard';
+import SearchBanner from '@/components/search-page/SearchBanner';
 import { photosExample } from '@/components/rec-resource/RecResourcePage';
 
 const SearchPage = () => {
